LPS-89691 Cast id to integer until generator generates id as long

+++ modules/apps/headless/headless-web-experience/headless-web-experience-impl/src/main/java/com/liferay/headless/web/experience/internal/resource/StructuredContentResourceImpl.java
@@ -86,7 +86,8 @@ public class StructuredContentResourceImpl
 			journalArticle -> {
 				StructuredContent structuredContent = new StructuredContent();
 
-				structuredContent.setId(journalArticle.getResourcePrimKey());
+				structuredContent.setId(
+					(int)journalArticle.getResourcePrimKey());
 
 				return structuredContent;
 			}
